organized type/classes
