+++ src/dynnn/types/enums.py
@@ -1,0 +1,32 @@
+from enum import Enum
+
+
+class GeneratorType(Enum):
+    LAGRANGIAN = 1
+    HAMILTONIAN = 2
+
+    @classmethod
+    def _missing_(cls, value):
+        try:
+            return cls[value.upper()]
+        except KeyError:
+            raise ValueError(f"{value} is not a valid GeneratorType")
+
+
+class OdeSolver(Enum):
+    # On choosing an ODE solver: https://docs.sciml.ai/DiffEqDocs/stable/solvers/ode_solve/
+    TSIT5 = 1
+    DOPRI5 = 2
+    ALF = 3
+    EULER = 4
+    MIDPOINT = 5
+    RK4 = 6
+    IEULER = 7
+    SYMPLECTIC = 8
+
+    @classmethod
+    def _missing_(cls, value):
+        try:
+            return cls[value.upper()]
+        except KeyError:
+            raise ValueError(f"{value} is not a valid OdeSolver")
